Long sequences and fast shuffling
LONG SEQUENCES
So far, the length of an input sequence could not exceed the size of an
int, which meant that the 3*10^9 nucleotides of the human genome could
not be specified using -l. Fixed.

FAST SHUFFLING
For cross-validation the input site frequency spectrum needs to be split
into two or more spectra. Prior to that, the input data are
shuffled. This shuffling step used to be carried out by calling the
function gsl_ran_shuffle from the GNU Scientific Library. Now it is
carried by shuffle in util.c, as my reimplementation is substantially
faster than the library routine.

+++ srcEpos/interface.c
@@ -66,7 +66,7 @@ Args *getArgs(int argc, char *argv[]){
     switch(c){
       break;
     case 'l':                           /* sequence length */
-      args->l = atoi(optarg);
+      args->l = atol(optarg);
       break;
     case 'L':                           /* preset levels */
       args->L = estrdup(optarg);
